+++ src/core/drd_dbus_constants.h
@@ -9,3 +9,6 @@
 
 #define DRD_LIGHTDM_REMOTE_FACTORY_BUS_NAME "org.deepin.DisplayManager"
 #define DRD_LIGHTDM_REMOTE_FACTORY_OBJECT_PATH  "/org/deepin/DisplayManager/RemoteDisplayFactory"
+
+#define DRD_LOGIND_BUS_NAME "org.freedesktop.login1"
+#define DRD_LOGIND_MANAGER_OBJECT_PATH "/org/freedesktop/login1"